reworked to use theme packages instead of manual
I add my own repo with a package that contains theming information and a
script to install defaults.

I've then gutted the original manual version, installing my packages and
running the scripts to set up the live theme.

This is more useful bc I will be able to use it on the installed system
as well instead of yet another manual set up.

Signed-off-by: Dylan Turner <dylantdmt@gmail.com>